requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
PyJWT==2.10.1
python-dotenv==1.0.1
Werkzeug==3.1.3
WTForms==3.2.1
